requirements.txt
accelerate==0.29.3
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
comtypes==1.4.15
confection==0.1.5
cryptography==46.0.4
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
decorator==4.4.2
deep-translator==1.11.4
distro==1.9.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
easyocr==1.7.2
edge-tts==7.2.7
elevenlabs==2.35.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gender-guesser==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.78.1
grpcio-status==1.71.2
gTTS==2.5.4
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.3
langdetect==1.0.9
lazy_loader==0.4
MarkupSafe==3.0.3
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
ninja==1.13.0
nltk==3.9.2
numpy==2.4.1
opencv-python-headless==4.13.0.92
packaging==26.0
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
poppler-utils==0.1.0
preshed==3.0.12
proglog==0.1.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyparsing==3.3.2
PyPDF2==3.0.1
pypdfium2==5.3.0
pypiwin32==223
pytesseract==0.3.13
python-bidi==0.6.7
python-dotenv==1.2.1
pyttsx3==2.99
pywin32==311
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.5
srsly==2.5.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
thinc==8.3.10
tifffile==2026.1.28
tokenizers==0.19.1
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.1
transformers==4.41.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
websockets==15.0.1
whitenoise==6.12.0
wrapt==2.0.1
yarl==1.22.0
